requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.1
frozenlist==1.4.1
future==1.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==9.0.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
typing_extensions==4.12.0
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.0
